Cleaned up training scripts
Training seems to be stable and more or less done for the first basic version. Removed references to other architectures and activation functions as we will not suppor this for now but might do so in the future. Saving and loading trained models is not implemented yet and is the next step.